matlabBindingGenerator code was not up to date with stable matlab generation code in cheader_reader.py. updating it

+++ matlabBindings/+helics/HELICS_INVALID_OPTION_INDEX.m
@@ -1,7 +1,7 @@
 function v = HELICS_INVALID_OPTION_INDEX()
 	persistent vInitialized;
 	if isempty(vInitialized)
-		vInitialized = -101;
+		vInitialized = int32(-101);
 	end
 	v = vInitialized;
 end
